requirements-backend.txt
Flask==2.3.2
Flask-CORS==4.0.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.0.0
numpy==1.24.3
matplotlib==3.7.2
tensorflow==2.13.0
opencv-python==4.8.0.76
